RAG-Questions-Creator-main/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2  
pypdf 
faiss-cpu
langchain-google-genai
langchain-community
langchain-core
cryptography
fpdf

